refactor(studio): remove Activity types from engine.ts, update 16 import sites, delete extracted domain files
Activity types and runtime functions removed from types/engine.ts (kept Bridge/IPC types). All 16 import sites updated to import from domain/activity shim. Deleted registry.ts, builtin.ts, operations.ts, circular-dependency.ts, circular-reference.ts, circular-dependency.test.ts.

Ultraworked with Sisyphus (https://github.com/code-yeongyu/oh-my-openagent)
